comments: don't hide comments made to lines outside the diff context
This is a hack but seems like the best way to solve the problem of users
increasing the context size and then making comments. The way things are done
makes it hard to come up with a better solution.

It should perhaps also be made impossible to comment on lines that isn't
changed in the changeset.

This issue is related to the need for a way to easily navigate to the
(unresolved) comments that has been made on a huge changeset.