SETTINGS_PREFIX for identifiers (e.g., db table names) incl. project's name.
kallithea.SETTINGS_PREFIX is a variable string used as a prefix for specific
external identifiers, such as database table names (and likely later form
fields), so that the name of the project need not necessarily be encoded into
data.

This setting is configurable so that compatibility with old, similar
databases can be maintained at the users' request.