comments: introduce lazy join of ChangesetStatus on ChangesetComment
99% of all uses of ChangesetComment will also need the optional corresponding
ChangesetStatus (if any) on the status_change relationship. Fetching it on
demand gives a lot of roundtrips and might be slow ... but adding explicit bulk
queries everywhere do not seem feasible. Adding lazy=joined in the data model
seems like the best solution.

Loading a pull request with 10 comments on a slow repo goes from 3.5 s to
2.7 s.