notifications: use stupid mail static-except-[] subjects to please gmail and its broken threading
This gives reasonable threading, both with gmail and proper mail clients, at
the cost of making the email subjects slightly obscure.

Gmail is clueless - also about mail threading, ignoring our References headers
and the wisdom of jwz. Gmail will start a new thread for each subject. The only
way to make all mails related to the same PR belong to the same thread is thus
to let all mails for a PR have the same subject. Gmail will however ignore
content in square brackets at the beginning of the subject so we can put the
interesting parts there.