helpers: drop h.tooltip
h.tooltip did more magic in the past - now it just did a douple (or triple?)
escape of html.

c9cfaeb1cdfe removed most of the need for double escaping - and the places
where it is used, it must be 'tagged' with the safe-html-title class. Thus,
none of the remaining uses of h.tooltip are relevant (and might even be wrong);
normal automatic template escaping is just fine.

This is mostly:
sed -i 's,title="\${h.tooltip(\([^}]*\))}",title="${\1}",g' `hg mani`