comments: limit comment deletion to a 5 minutes grace period
It is really a bad idea to delete a comment after the content and references to
it has been sent in mails.

Ultimately, comments should be editable and versioned - "delete" should then
just hide the comment.

Note: this is just done at the UI level - not verified on the server