pytest migration: use pytest test skipping instead of nose's
While the nose test skipping still worked, we want to remove all remnants of
nose, so switch to the pytest equivalent.

Note: due to a bug in pytest-sugar (reported at [1]) the skipped tests are
not shown explicitly when using the 'skipif' approach, while they were with
the nose approach. Since pytest-sugar is not enabled yet in setup.py, this
is not deemed a problem.

[1] https://github.com/Frozenball/pytest-sugar/issues/75