summary: add tooltip and link to status change icons (same as on changelog page)
This change makes the changeset list on the summary page more similar to the one
on the changeset list. It adds the same tooltip ("Changeset status") to the
changeset status change icons that is already found on the changelog page. It
also adds a link to the first comment (as in the changelog page).